Rename AddRemoveTabs to a more descriptive name PoolLiquidityHeader and move it to where it's used
