+++ src/components/Pool/PoolLiquidityHeader/index.tsx
@@ -1,0 +1,1 @@
+export { PoolLiquidityHeader } from './PoolLiquidityHeader'
