+++ src/pages/AddLiquidity/index.tsx
@@ -13,8 +13,8 @@ import { BlueCard, OutlineCard } from '../../components/Card'
 import { AutoColumn, ColumnCenter } from '../../components/Column'
 import { CurrencyInputPanel } from '../../components/CurrencyInputPanel'
 import DoubleCurrencyLogo from '../../components/DoubleLogo'
-import { AddRemoveTabs } from '../../components/NavigationTabs'
 import { PageMetaData } from '../../components/PageMetaData'
+import { PoolLiquidityHeader } from '../../components/Pool/PoolLiquidityHeader'
 import { MinimalPositionCard } from '../../components/PositionCard'
 import Row, { RowBetween, RowFlat } from '../../components/Row'
 import TradePrice from '../../components/Swap/TradePrice'
@@ -316,7 +316,7 @@ export default function AddLiquidity() {
     <>
       <PageMetaData title={`${isCreate ? 'Create a pair' : 'Add Liquidity'} | Swapr`} />
       <AppBody>
-        <AddRemoveTabs creating={isCreate} adding={true} />
+        <PoolLiquidityHeader creating={isCreate} adding={true} />
         <Wrapper>
           <TransactionConfirmationModal
             isOpen={showConfirm}
