+++ src/pages/RemoveLiquidity/index.tsx
@@ -16,9 +16,9 @@ import { AutoColumn, ColumnCenter } from '../../components/Column'
 import { CurrencyInputPanel } from '../../components/CurrencyInputPanel'
 import { CurrencyLogo } from '../../components/CurrencyLogo'
 import DoubleCurrencyLogo from '../../components/DoubleLogo'
-import { AddRemoveTabs } from '../../components/NavigationTabs'
 import { Option } from '../../components/Option'
 import { PageMetaData } from '../../components/PageMetaData'
+import { PoolLiquidityHeader } from '../../components/Pool/PoolLiquidityHeader'
 import { MinimalPositionCard } from '../../components/PositionCard'
 import Row, { RowBetween, RowFixed } from '../../components/Row'
 import Slider from '../../components/Slider'
@@ -501,7 +501,7 @@ export default function RemoveLiquidity() {
     <>
       <PageMetaData title="Liquidity | Swapr" />
       <AppBody>
-        <AddRemoveTabs creating={false} adding={false} />
+        <PoolLiquidityHeader creating={false} adding={false} />
         <Wrapper>
           <TransactionConfirmationModal
             isOpen={showConfirm}
